Changed some Card Effects
Cards that need to wait for a phase to end, like modifying roll ammount, will now be stagged for use when clicking the use button. Each phase will call a list of stagged cards at the aproate time to be used. thats when the cards effect will be applied. Other card can be used when button is pressed.